requirements.txt
speechRecognition
pyttsx3
pywhatkit
wikipedia
pyjokes
pyyaml


